Fix broken test
Also get sending errors roughed in
