+++ .eslintrc.json
@@ -6,8 +6,12 @@
   "extends": [
     "eslint:recommended",
     "plugin:@typescript-eslint/recommended",
+    "plugin:jest/recommended",
     "plugin:prettier/recommended",
     "prettier"
+  ],
+  "plugins": [
+    "jest"
   ],
   "parser": "@typescript-eslint/parser",
   "root": true
