+++ package.json
@@ -21,12 +21,12 @@
     "@swc/cli": "0.1.57",
     "@swc/core": "1.3.8",
     "@swc/jest": "0.2.23",
-    "@types/jest": "29.1.2",
     "@types/node": "17.0.16",
     "@typescript-eslint/eslint-plugin": "5.40.0",
     "@typescript-eslint/parser": "5.40.0",
     "eslint": "8.25.0",
     "eslint-config-prettier": "8.5.0",
+    "eslint-plugin-jest": "27.1.2",
     "eslint-plugin-prettier": "4.2.1",
     "jest": "29.2.0",
     "nanoid": "3.3.4",
