+++ src/index.test.ts
@@ -1,3 +1,4 @@
+import { describe, test, expect } from "@jest/globals";
 import run from "./testing/run";
 import path from "path";
 import { nanoid } from "nanoid";
