+++ src/post-errors.test.ts
@@ -1,0 +1,7 @@
+import { describe, test, expect } from "@jest/globals";
+
+describe("postErrors", () => {
+  test("posts errors", async () => {
+    expect(true).toBe(true);
+  });
+});
